fix: full codebase audit — 45 bugs across 17 files
Memory subsystem:
- GC Phase 3: skip worker heaps to prevent race with post-task cleanup
- GC Phase 4: scan owner freelist/slabs + all foreign lists (read-only)
  for safe oversized pool reclamation without cross-heap mutation
- ray_pool_of: return NULL on fallback validation failure
- NULL guards on all 7 ray_pool_of call sites (flush_foreign, flush_slabs,
  return_foreign_freelist, heap_merge)
- Foreign-block free: stop charging bytes_allocated to wrong thread
- ray_scratch_realloc: skip ray_detach_owned_refs for arena blocks
- pool.c: add seq_cst fence before ray_rc_sync reset

Joins / sort / group:
- ASOF join: rewind right pointer on partition change
- ASOF join: type-safe eq key access via read_col_i64 (was int64_t* cast)
- Hash join: NULL keys no longer match (SQL NULL != NULL semantics)
- GROUP BY: cap radix phase2 init_grp at 65536
- Sort: unsigned arithmetic for composite radix F64 range subtraction

Optimizer:
- graph_alloc_node_opt: NULL guards on OP_WINDOW_JOIN pointer fixup
- graph_alloc_node_opt: add OP_ANTIJOIN and OP_PIVOT fixup cases
- DCE mark_live: add OP_ANTIJOIN and OP_PIVOT traversal
- fold_binary_const: add RAY_I32/DATE/TIME constant folding
- Type inference + constant fold: enlarge DFS stack to nc*2+64

Executor:
- parted_gather_col: guard against zero-length VLA (n_segs==0)
- sel_compact: skip RAY_MAPCOMMON columns
- exec_filter_vec: error on input/pred length mismatch, fix result leak
- expr.c: fix tautological || condition reading f64 over i64 atoms
- expr.c OP_CAST: add bounds check on register index

Language VM:
- op_callf: all validation before VM state mutation; stack limits
  checked before arity; fn_args/fn_obj released on all error paths
- op_callf: arity error includes "expected N args, got M" detail
- VM trap handler: prefer vm_err_obj over __raise_val (fixes leak)
- compile.c patch_jump: int16_t overflow check
- compile.c ray_compile_reset: reset sf_self
- env.c: increase ENV_CAP from 512 to 1024

String ops:
- exec_replace: save prev before ray_str_vec_append to fix OOM leak
- exec_concat: detect scalar null atoms via RAY_ATOM_IS_NULL
- exec_concat: save prev before ray_str_vec_append to fix OOM leak
- exec_string_unary: same OOM leak fix

Graph engine:
- Shortest path: return error for paths > 254 hops (was silent truncation)
- Dijkstra: reject negative edge weights before main loop

Storage / I/O:
- HNSW load: validate layer n_nodes/M_max from untrusted files
- IPC: cap uncomp_size at 256 MB, guard di==0 before un-delta
- part.c: accept integer-only partition directory names (both locations)

+++ src/ops/sort.c
@@ -1029,7 +1029,7 @@ void radix_encode_fn(void* arg, uint32_t wid, int64_t start, int64_t end) {
                 } else {
                     val = 0;
                 }
-                uint64_t part = (uint64_t)(val - c->mins[k]);
+                uint64_t part = (uint64_t)val - (uint64_t)c->mins[k];
                 if (c->descs[k]) part = (uint64_t)c->ranges[k] - part;
                 composite |= part << c->bit_shifts[k];
             }
@@ -1406,7 +1406,7 @@ static void fused_topn_fn(void* arg, uint32_t wid,
             } else {
                 val = 0;
             }
-            uint64_t part = (uint64_t)(val - c->mins[k]);
+            uint64_t part = (uint64_t)val - (uint64_t)c->mins[k];
             if (c->descs[k]) part = (uint64_t)c->ranges[k] - part;
             composite |= part << c->bit_shifts[k];
         }
